HW 08 bad work

+++ src/main/webapp/resources/js/topjava.common.js
@@ -2,11 +2,12 @@ var form;
 
 function makeEditable() {
     form = $('#detailsForm');
-    $(".delete").click(function () {
-        if (confirm('Are you sure?')) {
-            deleteRow($(this).attr("id"));
-        }
-    });
+    //hw08 2
+    // $(".delete").click(function () {
+    //     if (confirm('Are you sure?')) {
+    //         deleteRow($(this).attr("id"));
+    //     }
+    // });
 
     $(document).ajaxError(function (event, jqXHR, options, jsExc) {
         failNoty(jqXHR);
@@ -20,16 +21,28 @@ function add() {
     form.find(":input").val("");
     $("#editRow").modal();
 }
-
+//hw08 2
 function deleteRow(id) {
-    $.ajax({
-        url: ctx.ajaxUrl + id,
-        type: "DELETE"
-    }).done(function () {
-        updateTable();
-        successNoty("Deleted");
-    });
+    if (confirm('Are you sure?')) {
+        $.ajax({
+            url: ctx.ajaxUrl + id,
+            type: "DELETE"
+        }).done(function () {
+            updateTable();
+            successNoty("Deleted");
+        });
+    }
 }
+///master version
+// function deleteRow(id) {
+//     $.ajax({
+//         url: ctx.ajaxUrl + id,
+//         type: "DELETE"
+//     }).done(function () {
+//         updateTable();
+//         successNoty("Deleted");
+//     });
+// }
 
 function updateTable() {
     $.get(ctx.ajaxUrl, function (data) {
